+++ SETOCryptomatorCryptor/SETOCryptor.h
@@ -21,7 +21,8 @@ typedef NS_ENUM(NSInteger, SETOCryptorError) {
 typedef NS_ENUM(NSInteger, SETOCryptorVersion) {
 	SETOCryptorVersion3 = 3,
 	SETOCryptorVersion4 = 4,
-	SETOCryptorVersion5 = 5
+	SETOCryptorVersion5 = 5,
+	SETOCryptorVersion6 = 6
 };
 
 typedef void (^SETOCryptorCompletionCallback)(NSError *error);