+++ SETOCryptomatorCryptor/SETOCryptorV3.h
@@ -8,6 +8,9 @@
 
 #import "SETOCryptor.h"
 
+/**
+ *  Vault format v3 was introduced in the first final release of Cryptomator 1.0.0. Older vault formats aren't supported anymore.
+ */
 @interface SETOCryptorV3 : SETOCryptor
 
 @end