+++ SETOCryptomatorCryptor/SETOCryptorV4.h
@@ -1,0 +1,16 @@
+//
+//  SETOCryptorV4.h
+//  SETOCryptomatorCryptor
+//
+//  Created by Tobias Hagemann on 19.06.17.
+//  Copyright © 2017 setoLabs. All rights reserved.
+//
+
+#import "SETOCryptorV3.h"
+
+/**
+ *  Beginning with vault format v4, directories have a 0 (zero) prefix instead of a _ (underscore) suffix. No implementation needed because this change isn't part of this library.
+ */
+@interface SETOCryptorV4 : SETOCryptorV3
+
+@end