+++ SETOCryptomatorCryptor/SETOCryptorV4.m
@@ -1,0 +1,13 @@
+//
+//  SETOCryptorV4.m
+//  SETOCryptomatorCryptor
+//
+//  Created by Tobias Hagemann on 19.06.17.
+//  Copyright © 2017 setoLabs. All rights reserved.
+//
+
+#import "SETOCryptorV4.h"
+
+@implementation SETOCryptorV4
+
+@end