+++ SETOCryptomatorCryptor/SETOCryptorV5.h
@@ -6,8 +6,11 @@
 //  Copyright © 2015-2016 setoLabs. All rights reserved.
 //
 
-#import "SETOCryptor.h"
+#import "SETOCryptorV4.h"
 
-@interface SETOCryptorV5 : SETOCryptor
+/**
+ *  Beginning with vault format v5, file size obfuscation has been disabled. File sizes can be determined in O(1) instead of having to read and decrypt the file header. This allows showing file sizes in the directory listing without having to download each file first. The file size in the header is now unused and filled with 0xFFFFFFFFFFFFFFFF.
+ */
+@interface SETOCryptorV5 : SETOCryptorV4
 
 @end